requirements/base.txt
Django==5.1.6
djangorestframework
python-dotenv